update.py: support multiple definitions per line
First, we parse all definitions using ctags. Then we tokenize all files
and assume token that match a definition are references.

For that to work, we must have a way to know if a given ident is defined
in a given (blob, line_number). defs_idxes is used for that: it maps
(idx*1000000 + line_number) to the ident defined on the line. This
allows O(1) lookup time.

This however assumes a single definition per line. See for example enum
EPOLL_EVENTS in musl v0.5.0 for a line that defines two idents. The DB
contains a definition AND a reference for this ident, at this line:

   enum EPOLL_EVENTS { __EPOLL_DUMMY };

https://elixir.bootlin.com/musl/v0.5.0/A/ident/EPOLL_EVENTS

Replace old dict by a set containing (idx, line_number, ident) tuples.
This allows storing multiple definitions for the same line. This has
the benefit of simplifying the code, as the testing condition becomes
more straight forward.

Signed-off-by: Théo Lebrun <theo.lebrun@bootlin.com>