README: drop "Hardware requirements" section
We know of no other big Elixir instance. Giving recommandations sounds
useless, especially as they are out of date.

If you ever read this: the recommendation is to try it out! Indexing is
expensive, but once done the front-end is easy to run. Put the two on
two different machines might be a good solution.

On prod (see prod branch), we do a A/B mechanism so that updating
happens on B while HTTP is served from A. That also means the Docker
contains mounts data read-only.

Signed-off-by: Théo Lebrun <theo.lebrun@bootlin.com>