README: drop "Using a cache to improve performance" section
I am not sure we are still running it. Anyway, if we do, last time we
checked its hit rate was super low. This is not a surprise: we serve
millions (billions?) of different pages, and we have wandering bots
that fill up the cache with odd pages.

We wouldn't recommend running a cache, instead sizing up Elixir
(and contributing perf improvements) is a much better strategy.

Signed-off-by: Théo Lebrun <theo.lebrun@bootlin.com>